Patch a potential CLR space leak and improve perf. by 5-10x
We generate a delegate for each CLR call and never release them,
building up a space leak.

This change leverages unsafePerformIO to cache the function pointers and
the delegates, restricting the space leak to one per inline, not one per call,
and in doing so improves performance by a factor of 10x
(less noticeable when (un)marshalling is involved).